+++ test/core/always_prefix.sv
@@ -1,0 +1,2 @@
+`define ALWAYS(trigger) always_comb
+`include "always_prefix.vh"