+++ test/core/always_prefix.v
@@ -1,0 +1,2 @@
+`define ALWAYS(trigger) always @(trigger)
+`include "always_prefix.vh"